+++ packages/template-retail-react-app/app/components/product-view/index.jsx
@@ -39,7 +39,7 @@ import {API_ERROR_MESSAGE} from '@salesforce/retail-react-app/app/constants'
 import DisplayPrice from '@salesforce/retail-react-app/app/components/display-price'
 import {getDisplayPrice} from '@salesforce/retail-react-app/app/utils/product-utils'
 
-const ProductViewHeader = ({name, basePrice, discountPrice, currency, category, productType}) => {
+const ProductViewHeader = ({name, basePrice, discountPrice, isProductPartOfBundle, currency, category, productType}) => {
     const isProductASet = productType?.set
 
     return (
@@ -55,12 +55,14 @@ const ProductViewHeader = ({name, basePrice, discountPrice, currency, category, 
                 <Heading fontSize="2xl">{`${name}`}</Heading>
             </Skeleton>
 
-            <DisplayPrice
-                basePrice={basePrice}
-                discountPrice={discountPrice}
-                currency={currency}
-                isProductASet={isProductASet}
-            />
+            {!isProductPartOfBundle && (
+                <DisplayPrice
+                    basePrice={basePrice}
+                    discountPrice={discountPrice}
+                    currency={currency}
+                    isProductASet={isProductASet}
+                />
+            )}
         </VStack>
     )
 }
@@ -71,7 +73,8 @@ ProductViewHeader.propTypes = {
     discountPrice: PropTypes.number,
     currency: PropTypes.string,
     category: PropTypes.array,
-    productType: PropTypes.object
+    productType: PropTypes.object,
+    isProductPartOfBundle: PropTypes.bool
 }
 
 const ButtonWithRegistration = withRegistration(Button)
@@ -95,10 +98,15 @@ const ProductView = forwardRef(
             updateWishlist,
             isProductLoading,
             isProductPartOfSet = false,
+            isProductPartOfBundle = false,
+            childOfBundleQuantity = 0,
+            childProductOrderability,
+            setChildProductOrderability,
             isBasketLoading = false,
             onVariantSelected = () => {},
             validateOrderability = (variant, quantity, stockLevel) =>
-                !isProductLoading && variant?.orderable && quantity > 0 && quantity <= stockLevel
+                !isProductLoading && variant?.orderable && quantity > 0 && quantity <= stockLevel,
+            showImageGallery = true
         },
         ref
     ) => {
@@ -125,17 +133,18 @@ const ProductView = forwardRef(
             variationAttributes,
             stockLevel,
             stepQuantity
-        } = useDerivedProduct(product, isProductPartOfSet)
+        } = useDerivedProduct(product, isProductPartOfSet, isProductPartOfBundle)
         const {basePrice, discountPrice} = getDisplayPrice(product)
         const canAddToWishlist = !isProductLoading
         const isProductASet = product?.type.set
+        const isProductABundle = product?.type.bundle
         const errorContainerRef = useRef(null)
 
         const validateAndShowError = (opts = {}) => {
             const {scrollErrorIntoView = true} = opts
             // Validate that all attributes are selected before proceeding.
             const hasValidSelection = validateOrderability(variant, quantity, stockLevel)
-            const showError = !isProductASet && !hasValidSelection
+            const showError = !isProductASet && !isProductABundle && !hasValidSelection
             const scrollToError = showError && scrollErrorIntoView
 
             toggleShowOptionsMessage(showError)
@@ -165,6 +174,10 @@ const ProductView = forwardRef(
                     defaultMessage: 'Add Set to Cart',
                     id: 'product_view.button.add_set_to_cart'
                 }),
+                addBundleToCart: intl.formatMessage({
+                    defaultMessage: 'Add Bundle to Cart',
+                    id: 'product_view.button.add_bundle_to_cart'
+                }),
                 addToWishlist: intl.formatMessage({
                     defaultMessage: 'Add to Wishlist',
                     id: 'product_view.button.add_to_wishlist'
@@ -172,6 +185,10 @@ const ProductView = forwardRef(
                 addSetToWishlist: intl.formatMessage({
                     defaultMessage: 'Add Set to Wishlist',
                     id: 'product_view.button.add_set_to_wishlist'
+                }),
+                addBundleToWishlist: intl.formatMessage({
+                    defaultMessage: 'Add Bundle to Wishlist',
+                    id: 'product_view.button.add_bundle_to_wishlist'
                 })
             }
 
@@ -191,7 +208,7 @@ const ProductView = forwardRef(
 
                 if (!addToCart && !updateCart) return null
                 if (updateCart) {
-                    await updateCart(variant, quantity)
+                    await updateCart(variant || product, quantity)
                     return
                 }
                 try {
@@ -202,7 +219,8 @@ const ProductView = forwardRef(
                     if (itemsAdded) {
                         onAddToCartModalOpen({
                             product,
-                            itemsAdded
+                            itemsAdded,
+                            selectedQuantity: quantity
                         })
                     }
                 } catch (e) {
@@ -219,12 +237,21 @@ const ProductView = forwardRef(
                 addToWishlist(product, variant, quantity)
             }
 
-            if (addToCart || updateCart) {
+            let disableButton = showInventoryMessage
+            if ((isProductASet || isProductABundle) && childProductOrderability) {
+                // if any of the children are not orderable, it will disable the add to cart button
+                disableButton = Object.keys(childProductOrderability).some((productId) => {
+                    return !childProductOrderability[productId]
+                })
+            }
+
+            // child product of bundles do not have add to cart button
+            if ((addToCart || updateCart) && !isProductPartOfBundle) {
                 buttons.push(
                     <Button
                         key="cart-button"
                         onClick={handleCartItem}
-                        disabled={isBasketLoading || showInventoryMessage}
+                        isDisabled={disableButton}
                         isLoading={isBasketLoading}
                         width="100%"
                         variant="solid"
@@ -234,12 +261,15 @@ const ProductView = forwardRef(
                             ? buttonText.update
                             : isProductASet
                             ? buttonText.addSetToCart
+                            : isProductABundle
+                            ? buttonText.addBundleToCart
                             : buttonText.addToCart}
                     </Button>
                 )
             }
 
-            if (addToWishlist || updateWishlist) {
+            // child product of bundles do not have add to wishlist button
+            if ((addToWishlist || updateWishlist) && !isProductPartOfBundle) {
                 buttons.push(
                     <ButtonWithRegistration
                         key="wishlist-button"
@@ -254,6 +284,8 @@ const ProductView = forwardRef(
                             ? buttonText.update
                             : isProductASet
                             ? buttonText.addSetToWishlist
+                            : isProductABundle
+                            ? buttonText.addBundleToWishlist
                             : buttonText.addToWishlist}
                     </ButtonWithRegistration>
                 )
@@ -275,7 +307,11 @@ const ProductView = forwardRef(
         }, [location.pathname])
 
         useEffect(() => {
-            if (!isProductASet && validateOrderability(variant, quantity, stockLevel)) {
+            if (
+                !isProductASet &&
+                !isProductABundle &&
+                validateOrderability(variant, quantity, stockLevel)
+            ) {
                 toggleShowOptionsMessage(false)
             }
         }, [variationParams])
@@ -285,6 +321,16 @@ const ProductView = forwardRef(
                 onVariantSelected(product, variant, quantity)
             }
         }, [variant?.productId, quantity])
+
+        useEffect(() => {
+            if (isProductPartOfBundle || isProductPartOfSet) {
+                // when showInventoryMessage is true, it means child product is not orderable
+                setChildProductOrderability((previousState) => ({
+                    ...previousState,
+                    [product.id]: !showInventoryMessage
+                }))
+            }
+        }, [showInventoryMessage])
 
         return (
             <Flex direction={'column'} data-testid="product-view" ref={ref}>
@@ -297,35 +343,38 @@ const ProductView = forwardRef(
                         productType={product?.type}
                         currency={product?.currency}
                         category={category}
+                        isProductPartOfBundle={isProductPartOfBundle}
                     />
                 </Box>
                 <Flex direction={['column', 'column', 'column', 'row']}>
-                    <Box flex={1} mr={[0, 0, 0, 6, 6]}>
-                        {product ? (
-                            <>
-                                <ImageGallery
-                                    size={imageSize}
-                                    imageGroups={product.imageGroups}
-                                    selectedVariationAttributes={variationParams}
-                                    lazy={isProductPartOfSet}
-                                />
-                                <HideOnMobile>
-                                    {showFullLink && product && (
-                                        <Link to={`/product/${product.master.masterId}`}>
-                                            <Text color="blue.600">
-                                                {intl.formatMessage({
-                                                    defaultMessage: 'See full details',
-                                                    id: 'product_view.link.full_details'
-                                                })}
-                                            </Text>
-                                        </Link>
-                                    )}
-                                </HideOnMobile>
-                            </>
-                        ) : (
-                            <ImageGallerySkeleton />
-                        )}
-                    </Box>
+                    {showImageGallery && (
+                        <Box flex={1} mr={[0, 0, 0, 6, 6]}>
+                            {product ? (
+                                <>
+                                    <ImageGallery
+                                        size={imageSize}
+                                        imageGroups={product.imageGroups}
+                                        selectedVariationAttributes={variationParams}
+                                        lazy={isProductPartOfSet || isProductPartOfBundle}
+                                    />
+                                    <HideOnMobile>
+                                        {showFullLink && product && (
+                                            <Link to={`/product/${product.master?.masterId}`}>
+                                                <Text color="blue.600">
+                                                    {intl.formatMessage({
+                                                        defaultMessage: 'See full details',
+                                                        id: 'product_view.link.full_details'
+                                                    })}
+                                                </Text>
+                                            </Link>
+                                        )}
+                                    </HideOnMobile>
+                                </>
+                            ) : (
+                                <ImageGallerySkeleton />
+                            )}
+                        </Box>
+                    )}
 
                     {/* Variations & Quantity Selector & CTA buttons */}
                     <VStack align="stretch" spacing={8} flex={1}>
@@ -337,9 +386,23 @@ const ProductView = forwardRef(
                                 productType={product?.type}
                                 currency={product?.currency}
                                 category={category}
+                                isProductPartOfBundle={isProductPartOfBundle}
                             />
                         </Box>
                         <VStack align="stretch" spacing={4}>
+                            {isProductPartOfBundle && (
+                                <Box>
+                                    <Text fontWeight="medium" fontSize="md" aria-label="price">
+                                        <label>
+                                            {intl.formatMessage({
+                                                defaultMessage: 'Quantity',
+                                                id: 'product_view.label.quantity'
+                                            })}
+                                            : {childOfBundleQuantity}
+                                        </label>
+                                    </Text>
+                                </Box>
+                            )}
                             {/*
                                 Customize the skeletons shown for attributes to suit your needs. At the point
                                 that we show the skeleton we do not know how many variations are selectable. So choose
@@ -415,8 +478,7 @@ const ProductView = forwardRef(
                                 </>
                             )}
 
-                            {/* Quantity Selector */}
-                            {!isProductASet && (
+                            {!isProductASet && !isProductPartOfBundle && (
                                 <VStack align="stretch" maxWidth={'200px'}>
                                     <Box fontWeight="bold">
                                         <label htmlFor="quantity">
@@ -473,7 +535,7 @@ const ProductView = forwardRef(
                             </Box>
                             <HideOnDesktop>
                                 {showFullLink && product && (
-                                    <Link to={`/product/${product.master.masterId}`}>
+                                    <Link to={`/product/${product.master?.masterId}`}>
                                         <Text color="blue.600">
                                             {intl.formatMessage({
                                                 defaultMessage: 'See full details',
@@ -510,7 +572,11 @@ const ProductView = forwardRef(
                     position="fixed"
                     bg="white"
                     width="100%"
-                    display={isProductPartOfSet ? 'none' : ['block', 'block', 'block', 'none']}
+                    display={
+                        isProductPartOfSet || isProductPartOfBundle
+                            ? 'none'
+                            : ['block', 'block', 'block', 'none']
+                    }
                     p={[4, 4, 6]}
                     left={0}
                     bottom={0}
@@ -529,6 +595,8 @@ ProductView.displayName = 'ProductView'
 ProductView.propTypes = {
     product: PropTypes.object,
     isProductPartOfSet: PropTypes.bool,
+    isProductPartOfBundle: PropTypes.bool,
+    childOfBundleQuantity: PropTypes.number,
     category: PropTypes.array,
     isProductLoading: PropTypes.bool,
     isBasketLoading: PropTypes.bool,
@@ -539,8 +607,11 @@ ProductView.propTypes = {
     updateWishlist: PropTypes.func,
     showFullLink: PropTypes.bool,
     imageSize: PropTypes.oneOf(['sm', 'md']),
+    childProductOrderability: PropTypes.object,
+    setChildProductOrderability: PropTypes.func,
     onVariantSelected: PropTypes.func,
-    validateOrderability: PropTypes.func
+    validateOrderability: PropTypes.func,
+    showImageGallery: PropTypes.bool
 }
 
 export default ProductView