+++ packages/template-retail-react-app/app/pages/account/wishlist/partials/wishlist-primary-action.test.js
@@ -13,6 +13,7 @@ import {screen, waitFor} from '@testing-library/react'
 import PropTypes from 'prop-types'
 import {rest} from 'msw'
 import {basketWithProductSet} from '@salesforce/retail-react-app/app/pages/product-detail/index.mock'
+import {mockProductBundle} from '@salesforce/retail-react-app/app/mocks/product-bundle'
 
 const MockedComponent = ({variant}) => {
     return (
@@ -97,3 +98,10 @@ test('the View Options button', async () => {
         {timeout: 5000}
     )
 }, 30000)
+
+test('bundle in wishlist renders the View Full Details button', async () => {
+    renderWithProviders(<MockedComponent variant={mockProductBundle} />)
+
+    const link = await screen.findByText(/view full details/i)
+    expect(link).toBeInTheDocument()
+})